flask/requirements.txt
flask
mysql-connector-python
python-dotenv
gunicorn
psutil
pytest
